requirements.txt
# Core dependencies
kafka-python==2.0.2
confluent-kafka==2.3.0
schema-registry-client==2.5.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MCP Framework
mcp==0.1.0
fastapi==0.104.1
uvicorn==0.24.0

# Google ADK
google-adk==0.1.0  # Replace with actual version

# RAG Pipeline
chromadb==0.4.18
langchain==0.1.0
sentence-transformers==2.2.2
openai==1.6.1

# Monitoring
prometheus-client==0.19.0
aiohttp==3.9.1

# Security
apache-ranger==0.3.0  # Replace with actual client

# Utilities
python-dotenv==1.0.0
click==8.1.7
croniter==2.0.1
aiocron==1.8
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
